KAN-342-gd / Fix: Jpa<Workspace,Long -> Integer>

+++ src/main/java/GASB/register_management/repository/WorkspaceRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface WorkspaceRepository extends JpaRepository<Workspace, Long> {
+public interface WorkspaceRepository extends JpaRepository<Workspace, Integer> {
 
     List<Workspace> findByIdIn(List<Integer> configIds);
 }
